feat: hf search
